Cpappas/create ent page type (#108)
* Create EnterprisePage type during build

* Fix linting issue

* minor refactor of gatsby-node.js

fixing a parens

Adding env var for determining mock data to use

Adding some style changes

Updating filepath for programs page mock data

Refactoring createPagesWithData. Extracting logic into its own file to make testing easier

* Cleaning up some lint
