+++ .env.development
@@ -10,10 +10,12 @@ SEGMENT_KEY=''
 DESIGNER_BASE_URL='http://designer.app:18808'
 HOSTNAME='example'
 UNBRANDED_LANDING_PAGE=''
-USE_MOCK_DATA=''
+USE_PROGRAMS_MOCK_DATA=''
+USE_ENTERPRISE_MOCK_DATA=''
 IDP_SLUG='saml-default'
-# uncomment the line below to use mock designer data from plugins/gatsby-source-wagtail/test/mock.json
-# USE_MOCK_DATA=true
+# uncomment ONLY 1 of the lines below to use mock designer data from plugins/gatsby-source-wagtail/test/
+# USE_PROGRAMS_MOCK_DATA=true
+# USE_ENTERPRISE_MOCK_DATA=true
 # uncomment the line below to simulate a deployment in which there is no IDP
 # IDP_SLUG=''
 # uncomment the line below to host the site at /{HOSTNAME}/
